backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.7
authlib==1.3.0
PyJWT==2.8.0
msal==1.26.0
